+++ NEWS.md
@@ -1,3 +1,15 @@
+1.22.1 - 2015-05-14
+-------------------
+
+Bug fixes:
+
+ * Close stream responses on error (#219)
+
+Enhancements:
+
+ * Format files split into pg and ogr directories
+
+
 1.22.0 - 2015-04-09
 -------------------